requirements.txt
langgraph
langchain
langchain-google-genai
python-dotenv
tiktoken
beautifulsoup4
requests
streamlit